+++ CHANGELOG.md
@@ -13,6 +13,10 @@
   * Stream validation for URLs has been made more robust
   * Minor bugfixes with Airplay 1 stream handling
   * Reset API password, support tunnels during factory reset
+  * Rename Updater to Updater/Admin Control Panel
+  * Add "Admin Settings" tab to the Updater/Admin Control Panel
+    * Add toggleable option to persist system logs to Admin Settings
+    * Move "Set password" page to be accessible via Admin Settings tab
 Streams
   * Fixed a bug where streams would disappear during upgrades by disabling validation when loading from disk